move: 쿼리 이동
